2658-maximum-number-of-fish-in-a-grid

+++ 2658-maximum-number-of-fish-in-a-grid/2658-maximum-number-of-fish-in-a-grid.java
@@ -6,7 +6,7 @@ class Solution {
         for (int i = 0; i < m; i++) {
             for (int j = 0; j < n; j++) {
                 if (grid[i][j] != 0) {
-                    long valueOfIsland = bfs(grid, i, j);
+                    int valueOfIsland = bfs(grid, i, j);
                     res = Math.max(res, valueOfIsland);
                 }
             }
